+++ Network/API/Telegram/Bot/Object/Update/Callback/Notification.hs
@@ -8,7 +8,7 @@ import Network.API.Telegram.Bot.Property.Persistable (Persistable (Payload, payl
 
 data Notification
 
-instance Persistable 'Post Notification where
-	type instance Payload 'Post Notification = Tagged ('Post Notification) (Text, Text)
-	payload (untag -> (cbq_id, text)) = object ["callback_query_id" .= cbq_id, "text" .= text]
-	endpoint _ = "answerCallbackQuery"
+-- instance Persistable 'Post Notification where
+-- 	type instance Payload 'Post Notification = Tagged ('Post Notification) (Text, Text)
+-- 	payload (untag -> (cbq_id, text)) = object ["callback_query_id" .= cbq_id, "text" .= text]
+-- 	endpoint _ = "answerCallbackQuery"