+++ Network/API/Telegram/Bot/Object/Update/Message/Content/Poll.hs
@@ -32,34 +32,34 @@ instance FromJSON Poll where
 		state :: Object -> State -> Parser Poll
 		state v f = f <$> v .: "id" <*> v .: "question" <*> v .: "options"
 
-instance Persistable ('Send 'Notify 'Directly) Poll where
-	type instance Payload ('Send 'Notify 'Directly) Poll
-		= Tagged ('Send 'Notify 'Directly Poll) (Int64, Text, [Text])
-	payload (untag -> (chat_id, question, options)) =
-		object ["chat_id" .= chat_id, "question" .= question,
-			"options" .= options, "disable_notification" .= False]
-	endpoint _ = "sendPoll"
-
-instance Persistable ('Send 'Silently 'Directly) Poll where
-	type instance Payload ('Send 'Silently 'Directly) Poll
-		= Tagged ('Send 'Silently 'Directly Poll) (Int64, Text, [Text])
-	payload (untag -> (chat_id, question, options)) =
-		object ["chat_id" .= chat_id, "question" .= question,
-			"options" .= options, "disable_notification" .= True]
-	endpoint _ = "sendPoll"
-
-instance Persistable ('Send 'Notify 'Replying) Poll where
-	type instance Payload ('Send 'Notify 'Replying) Poll
-		= Tagged ('Send 'Notify 'Replying Poll) (Int64, Int, Text, [Text])
-	payload (untag -> (chat_id, reply_to_message_id, question, options)) =
-		object ["chat_id" .= chat_id, "question" .= question, "options" .= options,
-			 "reply_to_message_id" .= reply_to_message_id, "disable_notification" .= False]
-	endpoint _ = "sendPoll"
-
-instance Persistable ('Send 'Silently 'Replying) Poll where
-	type instance Payload ('Send 'Silently 'Replying) Poll
-		= Tagged ('Send 'Silently 'Replying Poll) (Int64, Int, Text, [Text])
-	payload (untag -> (chat_id, reply_to_message_id, question, options)) =
-		object ["chat_id" .= chat_id, "question" .= question, "options" .= options,
-			"reply_to_message_id" .= reply_to_message_id, "disable_notification" .= True]
-	endpoint _ = "sendPoll"
+-- instance Persistable ('Send 'Notify 'Directly) Poll where
+-- 	type instance Payload ('Send 'Notify 'Directly) Poll
+-- 		= Tagged ('Send 'Notify 'Directly Poll) (Int64, Text, [Text])
+-- 	payload (untag -> (chat_id, question, options)) =
+-- 		object ["chat_id" .= chat_id, "question" .= question,
+-- 			"options" .= options, "disable_notification" .= False]
+-- 	endpoint _ = "sendPoll"
+--
+-- instance Persistable ('Send 'Silently 'Directly) Poll where
+-- 	type instance Payload ('Send 'Silently 'Directly) Poll
+-- 		= Tagged ('Send 'Silently 'Directly Poll) (Int64, Text, [Text])
+-- 	payload (untag -> (chat_id, question, options)) =
+-- 		object ["chat_id" .= chat_id, "question" .= question,
+-- 			"options" .= options, "disable_notification" .= True]
+-- 	endpoint _ = "sendPoll"
+--
+-- instance Persistable ('Send 'Notify 'Replying) Poll where
+-- 	type instance Payload ('Send 'Notify 'Replying) Poll
+-- 		= Tagged ('Send 'Notify 'Replying Poll) (Int64, Int, Text, [Text])
+-- 	payload (untag -> (chat_id, reply_to_message_id, question, options)) =
+-- 		object ["chat_id" .= chat_id, "question" .= question, "options" .= options,
+-- 			 "reply_to_message_id" .= reply_to_message_id, "disable_notification" .= False]
+-- 	endpoint _ = "sendPoll"
+--
+-- instance Persistable ('Send 'Silently 'Replying) Poll where
+-- 	type instance Payload ('Send 'Silently 'Replying) Poll
+-- 		= Tagged ('Send 'Silently 'Replying Poll) (Int64, Int, Text, [Text])
+-- 	payload (untag -> (chat_id, reply_to_message_id, question, options)) =
+-- 		object ["chat_id" .= chat_id, "question" .= question, "options" .= options,
+-- 			"reply_to_message_id" .= reply_to_message_id, "disable_notification" .= True]
+-- 	endpoint _ = "sendPoll"